+++ africanus/averaging/dask.py
@@ -6,17 +6,18 @@ from __future__ import print_function
 
 from operator import getitem
 
-from africanus.averaging.time_and_channel_mapping import (row_mapper,
-                                                          channel_mapper)
-from africanus.averaging.time_and_channel_avg import (row_average,
-                                                      row_chan_average,
-                                                      chan_average,
-                                                      merge_flags,
-                                                      AVERAGING_DOCS,
-                                                      AverageOutput,
-                                                      ChannelAverageOutput,
-                                                      RowAverageOutput,
-                                                      RowChanAverageOutput)
+from africanus.averaging.time_and_channel_mapping import (
+                row_mapper as np_row_mapper,
+                channel_mapper as np_channel_mapper)
+from africanus.averaging.time_and_channel_avg import (
+                row_average as np_row_average,
+                row_chan_average as np_row_chan_average,
+                chan_average as np_chan_average,
+                merge_flags as np_merge_flags,
+                AVERAGING_DOCS,
+                AverageOutput, ChannelAverageOutput,
+                RowAverageOutput, RowChanAverageOutput)
+
 from africanus.compatibility import PY3
 from africanus.util.requirements import requires_optional
 
@@ -33,33 +34,43 @@ else:
     dask_import_error = None
 
 
-def _row_chan_metadata(arrays, chan_bin_size):
+def chan_metadata(row_chan_arrays, chan_arrays, chan_bin_size):
     """ Create dask array with channel metadata for each chunk channel """
-    for array in arrays:
-        if array is None:
-            continue
+    chan_chunks = None
 
-        # Create a dask channel mapping structure
-        name = "channel-mapper-" + tokenize(array.chunks[1], chan_bin_size)
-        layers = {(name, i): (channel_mapper, c, chan_bin_size)
-                  for i, c in enumerate(array.chunks[1])}
-        graph = HighLevelGraph.from_collections(name, layers, ())
-        chunks = (array.chunks[1],)
-        chan_mapper = da.Array(graph, name, chunks, dtype=np.object)
+    for array in row_chan_arrays:
+        if array is not None:
+            chan_chunks = array.chunks[1]
+            break
 
-        return chan_mapper
+    if chan_chunks is None:
+        for array in chan_arrays:
+            if array is not None:
+                chan_chunks = array.chunks[0]
+                break
 
-    return None
+    if chan_chunks is None:
+        return None
+
+    # Create a dask channel mapping structure
+    name = "channel-mapper-" + tokenize(chan_chunks, chan_bin_size)
+    layers = {(name, i): (np_channel_mapper, c, chan_bin_size)
+              for i, c in enumerate(chan_chunks)}
+    graph = HighLevelGraph.from_collections(name, layers, ())
+    chunks = (chan_chunks,)
+    chan_mapper = da.Array(graph, name, chunks, dtype=np.object)
+
+    return chan_mapper
 
 
-def _dask_row_mapper(time, interval, antenna1, antenna2,
-                     flag_row=None, time_bin_secs=1.0):
+def row_mapper(time, interval, antenna1, antenna2,
+               flag_row=None, time_bin_secs=1.0):
     """ Create a dask row mapping structure for each row chunk """
 
     # dask py2 doesn't understand meta
     kw = {'meta': np.empty((0,), dtype=np.object)} if PY3 else {}
 
-    return da.blockwise(row_mapper, ("row",),
+    return da.blockwise(np_row_mapper, ("row",),
                         time, ("row",),
                         interval, ("row",),
                         antenna1, ("row",),
@@ -92,16 +103,16 @@ def _getitem_row(avg, idx, array, dims):
 def _row_average_wrapper(row_meta, ant1, ant2, flag_row,
                          time_centroid, exposure, uvw,
                          weight, sigma):
-    return row_average(row_meta, ant1, ant2, flag_row,
-                       time_centroid, exposure,
-                       uvw[0] if uvw is not None else None,
-                       weight[0] if weight is not None else None,
-                       sigma[0] if sigma is not None else None)
+    return np_row_average(row_meta, ant1, ant2, flag_row,
+                          time_centroid, exposure,
+                          uvw[0] if uvw is not None else None,
+                          weight[0] if weight is not None else None,
+                          sigma[0] if sigma is not None else None)
 
 
-def _dask_row_average(row_meta, ant1, ant2, flag_row=None,
-                      time_centroid=None, exposure=None, uvw=None,
-                      weight=None, sigma=None):
+def row_average(row_meta, ant1, ant2, flag_row=None,
+                time_centroid=None, exposure=None, uvw=None,
+                weight=None, sigma=None):
     """ Average row-based dask arrays """
 
     rd = ("row",)
@@ -154,10 +165,10 @@ def _getitem_row_chan(avg, idx, dtype):
 _row_chan_avg_dims = ("row", "chan", "corr")
 
 
-def _dask_row_chan_average(row_meta, chan_meta, flag_row=None, weight=None,
-                           vis=None, flag=None,
-                           weight_spectrum=None, sigma_spectrum=None,
-                           chan_bin_size=1):
+def row_chan_average(row_meta, chan_meta, flag_row=None, weight=None,
+                     vis=None, flag=None,
+                     weight_spectrum=None, sigma_spectrum=None,
+                     chan_bin_size=1):
     """ Average (row,chan,corr)-based dask arrays """
 
     if chan_meta is None:
@@ -171,7 +182,7 @@ def _dask_row_chan_average(row_meta, chan_meta, flag_row=None, weight=None,
     }
 
     flag_row_dims = None if flag_row is None else ("row",)
-    weight_dims = None if weight is None else ("row",)
+    weight_dims = None if weight is None else ("row", "corr")
     vis_dims = None if vis is None else _row_chan_avg_dims
     flag_dims = None if flag is None else _row_chan_avg_dims
     ws_dims = None if weight_spectrum is None else _row_chan_avg_dims
@@ -182,7 +193,7 @@ def _dask_row_chan_average(row_meta, chan_meta, flag_row=None, weight=None,
     kw = ({"meta": np.empty((0,)*len(_row_chan_avg_dims), dtype=np.object)}
           if PY3 else {})
 
-    avg = da.blockwise(row_chan_average, _row_chan_avg_dims,
+    avg = da.blockwise(np_row_chan_average, _row_chan_avg_dims,
                        row_meta, ("row",),
                        chan_meta, ("chan",),
                        flag_row, flag_row_dims,
@@ -216,8 +227,8 @@ def _getitem_chan(avg, idx, dtype):
     return da.Array(graph, name, avg.chunks, dtype=dtype, **kw)
 
 
-def _dask_chan_average(chan_meta, chan_freq=None, chan_width=None,
-                       chan_bin_size=1):
+def chan_average(chan_meta, chan_freq=None, chan_width=None,
+                 effective_bw=None, resolution=None, chan_bin_size=1):
 
     if chan_meta is None:
         return ChannelAverageOutput(None, None)
@@ -227,36 +238,40 @@ def _dask_chan_average(chan_meta, chan_freq=None, chan_width=None,
     }
 
     kw = {"meta": np.empty((0,), dtype=np.object)} if PY3 else {}
+    cdim = ("chan",)
 
-    avg = da.blockwise(chan_average, ("chan",),
-                       chan_meta, ("chan",),
-                       chan_freq, None if chan_freq is None else ("chan",),
-                       chan_width, None if chan_width is None else ("chan",),
+    avg = da.blockwise(np_chan_average, cdim,
+                       chan_meta, cdim,
+                       chan_freq, None if chan_freq is None else cdim,
+                       chan_width, None if chan_width is None else cdim,
+                       effective_bw, None if effective_bw is None else cdim,
+                       resolution, None if resolution is None else cdim,
                        adjust_chunks=adjust_chunks,
                        dtype=np.object,
                        **kw)
 
     tuple_gets = (None if a is None else _getitem_chan(avg, i, a.dtype)
-                  for i, a in enumerate([chan_freq, chan_width]))
+                  for i, a in enumerate([chan_freq, chan_width,
+                                         effective_bw, resolution]))
 
     return ChannelAverageOutput(*tuple_gets)
 
 
-def _dask_merge_flags(flag_row, flag):
+def merge_flags(flag_row, flag):
     """ Perform flag merging on dask arrays """
     if flag_row is None and flag is not None:
-        return da.blockwise(merge_flags, "r",
+        return da.blockwise(np_merge_flags, "r",
                             flag_row, None,
                             flag, "rfc",
                             concatenate=True,
                             dtype=flag.dtype)
     elif flag_row is not None and flag is None:
-        return da.blockwise(merge_flags, "r",
+        return da.blockwise(np_merge_flags, "r",
                             flag_row, "r",
                             None, None,
                             dtype=flag_row.dtype)
     elif flag_row is not None and flag is not None:
-        return da.blockwise(merge_flags, "r",
+        return da.blockwise(np_merge_flags, "r",
                             flag_row, "r",
                             flag, "rfc",
                             concatenate=True,
@@ -270,44 +285,49 @@ def time_and_channel(time, interval, antenna1, antenna2,
                      time_centroid=None, exposure=None, flag_row=None,
                      uvw=None, weight=None, sigma=None,
                      chan_freq=None, chan_width=None,
+                     effective_bw=None, resolution=None,
                      vis=None, flag=None,
                      weight_spectrum=None, sigma_spectrum=None,
                      time_bin_secs=1.0, chan_bin_size=1):
 
     row_chan_arrays = (vis, flag, weight_spectrum, sigma_spectrum)
+    chan_arrays = (chan_freq, chan_width, effective_bw, resolution)
 
     # The flow of this function should match that of the numba
     # time_and_channel implementation
 
     # Merge flag_row and flag arrays
-    flag_row = _dask_merge_flags(flag_row, flag)
+    flag_row = merge_flags(flag_row, flag)
 
     # Generate row mapping metadata
-    row_meta = _dask_row_mapper(time, interval,
-                                antenna1, antenna2,
-                                flag_row=flag_row,
-                                time_bin_secs=time_bin_secs)
+    row_meta = row_mapper(time, interval,
+                          antenna1, antenna2,
+                          flag_row=flag_row,
+                          time_bin_secs=time_bin_secs)
 
     # Generate channel mapping metadata
-    chan_meta = _row_chan_metadata(row_chan_arrays, chan_bin_size)
+    chan_meta = chan_metadata(row_chan_arrays, chan_arrays, chan_bin_size)
 
     # Average row data
-    row_data = _dask_row_average(row_meta, antenna1, antenna2,
-                                 flag_row=flag_row,
-                                 time_centroid=time_centroid,
-                                 exposure=exposure, uvw=uvw,
-                                 weight=weight, sigma=sigma)
+    row_data = row_average(row_meta, antenna1, antenna2,
+                           flag_row=flag_row,
+                           time_centroid=time_centroid,
+                           exposure=exposure, uvw=uvw,
+                           weight=weight, sigma=sigma)
 
     # Average channel data
-    row_chan_data = _dask_row_chan_average(row_meta, chan_meta,
-                                           flag_row=flag_row, weight=weight,
-                                           vis=vis, flag=flag,
-                                           weight_spectrum=weight_spectrum,
-                                           sigma_spectrum=sigma_spectrum,
-                                           chan_bin_size=chan_bin_size)
+    row_chan_data = row_chan_average(row_meta, chan_meta,
+                                     flag_row=flag_row, weight=weight,
+                                     vis=vis, flag=flag,
+                                     weight_spectrum=weight_spectrum,
+                                     sigma_spectrum=sigma_spectrum,
+                                     chan_bin_size=chan_bin_size)
 
-    chan_data = _dask_chan_average(chan_meta, chan_freq=chan_freq,
-                                   chan_width=chan_width)
+    chan_data = chan_average(chan_meta,
+                             chan_freq=chan_freq,
+                             chan_width=chan_width,
+                             effective_bw=effective_bw,
+                             resolution=resolution)
 
     # Merge output tuples
     return AverageOutput(_getitem_row(row_meta, 1, time, ("row",)),
@@ -323,6 +343,8 @@ def time_and_channel(time, interval, antenna1, antenna2,
                          row_data.sigma,
                          chan_data.chan_freq,
                          chan_data.chan_width,
+                         chan_data.effective_bw,
+                         chan_data.resolution,
                          row_chan_data.vis,
                          row_chan_data.flag,
                          row_chan_data.weight_spectrum,