+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 0.1.9 (YYYY-MM-DD)
 ------------------
 
+* Fix weight indexing in averaging (:pr:`144`)
+* Support EFFECTIVE_BW and RESOLUTION in averaging (:pr:`144`)
 * Optimise predict_vis jones coherency summation (:pr:`143`)
 * Remove use of @wraps (:pr:`141`, :pr:`142`)
 * Set row chunks to nan in dask averaging code. (:pr:`139`)